Fix for DIRSTUDIO-616 (Entry Editor Manager does not ensure that the entry's attributes are initialized (loaded) before testing and opening editors).
New fix as the previous one introduced a regression (the opening of the editor was synchronous with the load of attributes of the entry in the UI thread which could lead to a non responding UI in case the load of attributes takes a while).
This fix uses an asynchronous job, separate from the UI thread which gives the control back to the UI as soon as the job is scheduled.

git-svn-id: https://svn.apache.org/repos/asf/directory/studio/trunk@899200 13f79535-47bb-0310-9956-ffa450edef68